+++ tests/testthat/test_1-4-0_weight2rows.R
@@ -15,3 +15,16 @@ test_that("weight includes 0", {
   expect_equal(nrow(weight2rows(DT, "y")), 10L)
 })
 
+
+test_that("Logical weight still works", {
+  # Just for coverage
+  library(data.table)
+  DT <- data.table(x = 1:10,
+                   ww = as.logical(1:10 %% 3L))
+  DT4 <- suppressWarnings(weight2rows(DT, "ww", rows.out = 4L))
+  expect_equal(nrow(DT4), 4L)
+  DT4_no_w <- suppressWarnings(weight2rows(DT, "ww", rows.out = 4L, discard_weight.var = TRUE))
+  expect_false("ww" %in% names(DT4_no_w))
+})
+
+